[fix] secure Ad update and delete
    - make get request for Ads accessible without authentication
    - delete remove userId from Ad data in response to get Ad Request for security issues
    - in Ad Delete and Update check if the user sending this request is also the owner of the Ad .

+++ src/main/java/com/betroc/controller/AdBaseController.java
@@ -1,13 +1,19 @@
 package com.betroc.controller;
 
 import com.betroc.model.Advertisement;
+import com.betroc.model.User;
 import com.betroc.payload.ApiResponse;
 import com.betroc.repository.AdvertisementBaseRepository;
+import com.betroc.repository.UserRepository;
+import com.betroc.security.UserPrincipal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,6 +26,9 @@ public abstract class AdBaseController <T extends Advertisement,W extends Advert
     @Autowired
     private W repository;
 
+    @Autowired
+    private UserRepository userRepository;
+
     @GetMapping
     //@Secured("ROLE_USER")
     public Page getAllAds(@PageableDefault(size = 10, sort = "id") Pageable pageable){
@@ -29,8 +38,21 @@ public abstract class AdBaseController <T extends Advertisement,W extends Advert
 
     @PostMapping
     public ResponseEntity<?> registerAd(T ad){
-        repository.save(ad);
-         return ResponseEntity.accepted().body(new ApiResponse(true,"sucess"));
+
+        //get user data from his Authentication to create a new annonce
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserPrincipal userPrincipal = (UserPrincipal) auth.getPrincipal();
+        Optional userOp = userRepository.findById(userPrincipal.getId());
+
+        //if the user exist witch is normal :) then add the ad
+        if(userOp.isPresent()){
+            User user = (User) userOp.get();
+            ad.setUser(user);
+            repository.save(ad);
+            return ResponseEntity.accepted().body(new ApiResponse(true,"sucess"));
+        }
+
+        return ResponseEntity.accepted().body(new ApiResponse(true,"failed"));
     }
 
     @PutMapping
@@ -41,9 +63,20 @@ public abstract class AdBaseController <T extends Advertisement,W extends Advert
         if (!exchangeAdOptional.isPresent())
             return ResponseEntity.notFound().build();
 
+        T adWithOldData = (T) exchangeAdOptional.get();
+
+        //check if the user sending the request is the owner of the ad
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserPrincipal userPrincipal = (UserPrincipal) auth.getPrincipal();
+        if(userPrincipal.getId() != adWithOldData.getUser().getId())
+             return new ResponseEntity(new ApiResponse(false, "you do not have the rights to change this Ad"),
+                     HttpStatus.BAD_REQUEST);
+
+        //set the user
+        ad.setUser(adWithOldData.getUser());
+
         //if there is an Ad with the same id then update
-        T exchangeAdWithOldData = (T) exchangeAdOptional.get();
-        ad.setCreationDate(exchangeAdWithOldData.getCreationDate());
+        ad.setCreationDate(adWithOldData.getCreationDate());
         ad.setModificationDate(new Date());
         repository.save(ad);
 
@@ -56,7 +89,26 @@ public abstract class AdBaseController <T extends Advertisement,W extends Advert
 
         if(!repository.existsById(id))
             return ResponseEntity.accepted().body(new ApiResponse(false,"failed no such Ad"));
-        //TODO delete only with id not the all ad
+
+
+        //get auth data
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserPrincipal userPrincipal = (UserPrincipal) auth.getPrincipal();
+
+        Optional adOp = repository.findById(id);
+
+        //check is the Ad exist
+        if(!adOp.isPresent())
+            return ResponseEntity.notFound().build();
+
+
+        T ad = (T) adOp.get();
+
+        //check if the user sending the request is the owner of the ad
+        if (ad.getUser().getId() != userPrincipal.getId())
+            return new ResponseEntity(new ApiResponse(false, "you do not have the rights to change this Ad"),
+                    HttpStatus.BAD_REQUEST);
+
 
         repository.deleteById(id);
